+++ adapters/lumi/WXKG11LM.py
@@ -15,6 +15,7 @@ class WXKG11LM(AdapterWithBattery):
         self.switch.add_level('Long Click', 'hold')
         self.switch.add_level('Long Click Release', 'release')
         self.switch.set_selector_style(SelectorSwitch.SELECTOR_TYPE_MENU)
+        self.switch.disable_value_check_on_update()
 
         self.devices.append(self.switch)