+++ devices/boolean_sensor.py
@@ -19,10 +19,6 @@ class BooleanSensor(Device):
 
         return Domoticz.Device(Unit=unit, DeviceID=device_id, Name=device_name, Type=244, Subtype=73, Switchtype=self.sensor_type).Create()
 
-    def update_device(self, device, values):
-        if values['nValue'] != device.nValue or values['sValue'] != device.sValue:
-            super().update_device(device, values)
-
     def get_numeric_value(self, value, device):
         return 1 if value else 0