+++ devices/device.py
@@ -9,6 +9,7 @@ class Device():
         self.alias = alias
         self.value_key = value_key
         self.device_name_suffix = device_name_suffix
+        self.check_values_on_update = True
 
         if len(self.alias) > self.MAX_ALIAS_LENGTH:
             raise ValueError('Alias "' + self.alias + '" is too long to generate valid DeviceID')
@@ -45,8 +46,25 @@ class Device():
             '" value for device "' + device_name + '"'
         )
 
+    def disable_value_check_on_update(self):
+        self.check_values_on_update = False
+
     def update_device(self, device, values):
-        device.Update(**values)
+        nValueChanged = values['nValue'] != device.nValue
+        sValueChanged = values['sValue'] != device.sValue
+        colorChanged = 'Color' in values and values['Color'] != device.Color
+
+        if nValueChanged or sValueChanged or colorChanged or self.check_values_on_update == False:
+            device.Update(**values)
+        else:
+            self.touch_device(device)
+
+    def touch_device(self, device):
+        # Touch has been added in recent Domoticz beta, so check if it exists for backward compatibility
+        if hasattr(device, 'Touch') and callable(getattr(device, 'Touch')):
+            device.Touch()
+        else:
+            Domoticz.Debug('Received heartbeat message from device "' + device.Name + '"')
 
     # Register device in Domoticz
     def register(self, device_data):
@@ -92,8 +110,7 @@ class Device():
             return self._create_device(device_data)
 
         if (value == None):
-            # There is no way to properly handle heartbeat messages as nValue and sValue are mandatory for device update
-            Domoticz.Debug('Received heartbeat message from device "' + device.Name + '"')
+            self.touch_device(device)
             return None
 
         device_values = dict({